Keep tests for public mode and for aelo mode in 2 separate files
